refactor(trace-log): replace TraceLogService with TraceLogRepository and update logging methods for message engagement events
