+++ libs/application-generic/src/services/analytic-logs/index.ts
@@ -1,5 +1,4 @@
 export { RequestLogRepository, requestLogSchema, RequestLog } from './request-log';
-export { TraceLogRepository, traceLogSchema, Trace } from './trace-log';
-export * from './trace-log.service';
+export { TraceLogRepository, traceLogSchema, Trace, TraceEvent } from './trace-log';
 export * from './clickhouse.service';
 export * from './base.repository';
